+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+- Use garbage collection by default on all systems
 - Use `enzime` as the default user on `echo`
 - Removed `diff.age.textconv`
 - Use `/mnt/phi` on all `personal` NixOS systems